Fixed sorting issues
Fixed an issue where ignored rounds would still populate roundList when out of order in the retrieved JSON.
roundList is now sorted before the bracket is populated, just in case. It is assumed that IDs that increase in value as you traverse a round of sets from top to bottom.